requirements.txt
dj-database-url==0.4.2
Django==2.0.1
django-extensions==2.0.6
djangorestframework==3.8.2
Pillow==5.2.0
pytz==2018.5
six==1.11.0
